Adds product variation functionality
Adds the ability to select product variations during the addToCart process.

This includes changes to the cart, cart article, and billing components to handle and display product variations. It enhances the user experience by allowing customers to choose specific variations of a product, such as color or size, before adding it to their cart.

The customer model is also updated to check for the existence of previous orders.